test(example): reorganize react-router test layout
