+++ apps/react-router-example/project.json
@@ -17,7 +17,7 @@
       "executor": "nx:run-commands",
       "dependsOn": ["typegen", "prisma:generate", "build"],
       "options": {
-        "command": "node test/browser/run-browser-suite.mjs",
+        "command": "node tests/browser/support/run-browser-suite.mjs",
         "cwd": "apps/react-router-example"
       }
     },
