+++ apps/react-router-example/tests/browser/support/resolve-chrome-bin.ts
