+++ apps/react-router-example/tests/routes/_shared/effect-test.ts
@@ -1,0 +1,4 @@
+import * as Effect from "effect/Effect"
+
+export const runTestEffect = <A, E>(effect: Effect.Effect<A, E, unknown>) =>
+  Effect.runPromise(effect as Effect.Effect<A, E, never>)
